feat: data cleaning py script added
The script loads data from the otodom-data.xlsx file and processes it into the resulting cleaned-data.xlsx file. The script divides the location into individual parts. It classifies the type of seller, in addition, there is a conversion of field types of the outgoing Excel file and specifying the currency of the offer.